Moved errors to appropriate project structure - domain
