+++ src/Common/Common.Domain/Errors/ErrorCodes.cs
@@ -1,10 +1,14 @@
-﻿namespace Common.Presentation.Exceptions;
+﻿namespace Common.Domain.Errors;
 
+/// <summary>
+/// It's a partial class because it could be
+/// expanded in distinct services for their custom error codes
+/// </summary>
 public static partial class ErrorCodes
 {
     public const string DefaultError = "error";
 
-    public const string ValidationException = "validation_exception";
+    public const string Validation = "validation";
 
     public const string ServiceUnavailable = "service_unavailable";
 
@@ -13,6 +17,10 @@ public static partial class ErrorCodes
     public const string NotFound = "not_found";
 
     public const string Conflict = "conflict";
+
+    public const string Failure = "failure";
+
+    public const string Unexpected = "unexpected";
 
     public const string Forbidden = "forbidden";
 }
