+++ src/Common/Common.Domain/Errors/ErrorType.cs
@@ -1,4 +1,4 @@
-namespace Common.Presentation.Exceptions;
+namespace Common.Domain.Errors;
 
 public enum ErrorType
 {
