+++ src/Common/Common.Domain/Errors/ServiceException.cs
@@ -1,4 +1,4 @@
-﻿namespace Common.Presentation.Exceptions;
+﻿namespace Common.Domain.Errors;
 
 public class ServiceException : Exception
 {
